Fixed login and register

+++ app/src/main/java/com/example/a38096/fitnessproject/model/IWorkoutDataSource.java
@@ -17,7 +17,7 @@ public interface IWorkoutDataSource {
                                            int calories,
                                            int distance,
                                            int duration,
-                                           int workoutDate);
+                                           long workoutDate);
 
     Observable<ResponseBody> updateWorkout(String uuid,
                                            int workoutId,
